Updated item deletion confirmation form

+++ staff-member-area/deleteitem.php
@@ -6,30 +6,34 @@ require '../templates/project_footer.php';
 require '../functions/userauthorisation.php';
 require '../functions/finditembyid.php';
 require '../functions/validateimage.php';
+require '../functions/finduser.php';
 title_bar("Stock");
+print_r($_SESSION);
 if ($_SERVER['REQUEST_METHOD'] === 'POST') {
     $verify = findUserWithPwd($pdo, $_SESSION['email'], $_POST['password']);
-    if (isset($_GET["itemid"])) {
-        $item = findItemByID($pdo, $_GET['itemid']);
-        $sql = "DELETE from items WHERE id=?";
-        $stmt = $pdo->prepare($sql);
-        $stmt->execute([$item['id']]);
-        $sql = "DELETE FROM images WHERE filename =:filename";
-        $stmt = $pdo->prepare($sql);
-        $stmt->execute([":filename" => $item['image']]);
-        if ($item['image'] && !empty($item['image'])) {
-            $filePath = "../uploaded_images/" . $item['image'];
+    if ($verify) {
+        if (isset($_GET["itemid"])) {
+            $item = findItemByID($pdo, $_GET['itemid']);
+            $sql = "DELETE from items WHERE id=?";
+            $stmt = $pdo->prepare($sql);
+            $stmt->execute([$item['id']]);
+            $sql = "DELETE FROM images WHERE filename =:filename";
+            $stmt = $pdo->prepare($sql);
+            $stmt->execute([":filename" => $item['image']]);
+            if ($item['image'] && !empty($item['image'])) {
+                $filePath = "../uploaded_images/" . $item['image'];
 
 
-            if (file_exists($filePath)) {
-                unlink($filePath);
+                if (file_exists($filePath)) {
+                    unlink($filePath);
+                }
+            } else {
+                echo "<script>alert('Item deleted')</script>";
+                echo "<script>document.location = 'stock' </script>";
             }
-        } else {
-            echo "<script>alert('Item deleted')</script>";
+            echo "<script>alert('Item deleted " . $item['image'] . "')</script>";
             echo "<script>document.location = 'stock' </script>";
         }
-        echo "<script>alert('Item deleted " . $item['image'] . "')</script>";
-        echo "<script>document.location = 'stock' </script>";
     }
 }
 ?>
@@ -52,9 +56,10 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
             <button type="button" class="btn-continue">Yes</button>
         </div>
         <div class="hidden-pwd-box">
-            <label for="password">Password:</label>
-            <input type="password" name="password" required placeholder="please enter the password">
-            <button type="submit">Confirm Delete</button>
+            <label for="pwd-confirm-input">Password:</label>
+            <input type="password" name="password" id="pwd-confirm-input" required placeholder="please enter the password">
+            <button type="submit" class="confirm-buttons">Confirm Delete</button>
+            <button type="button" class="btn-cancel">Cancel</button>
         </div>
 
     </form>
